Rename ComponentMetaData rules
