--- build-conventions/src/main/java/org/elasticsearch/gradle/internal/DependencyContext.java
@@ -1,22 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-package org.elasticsearch.gradle.internal;
-
-import org.gradle.api.Named;
-import org.gradle.api.attributes.Attribute;
-
-public interface DependencyContext extends Named {
-    Attribute<DependencyContext> CONTEXT_ATTRIBUTE = Attribute.of("org.elasticsearch.context", DependencyContext.class);
-    String CODE_QUALITY = "code-quality";
-
-    default String getName() {
-        return "org.elasticsearch.context";
-    }
-}
