+++ build-tools-internal/src/main/java/org/elasticsearch/gradle/internal/dependencies/rules/ExcludeAllTransitivesRule.java
@@ -13,15 +13,30 @@ import org.gradle.api.artifacts.CacheableRule;
 import org.gradle.api.artifacts.ComponentMetadataContext;
 import org.gradle.api.artifacts.ComponentMetadataRule;
 
+/**
+ * A Gradle component metadata rule that excludes all transitive dependencies.
+ *
+ * <p>The rule operates on all variants of a component.</p>
+ *
+ * <h3>Example Usage:</h3>
+ * <pre>{@code
+ * dependencies {
+ *     components {
+ *        withModule("com.azure:azure-core-http-netty"", ExcludeAllTransitivesRule)
+ *     }
+ * }
+ * }</pre>
+ *
+ */
 @CacheableRule
-public abstract class Log4jRule implements ComponentMetadataRule {
+public abstract class ExcludeAllTransitivesRule implements ComponentMetadataRule {
+
     @Override
     public void execute(ComponentMetadataContext context) {
-        context.getDetails().withVariant("apiElements", variantMetadata -> {
-            variantMetadata.withDependencies(deps -> {
-                deps.add("org.jspecify:jspecify:1.0.0");
-                deps.add("com.github.spotbugs:spotbugs-annotations:4.9.3");
-                deps.add("com.google.errorprone:error_prone_annotations:2.38.0");
+        context.getDetails().allVariants(variant -> {
+            variant.withDependencies(dependencies -> {
+                // Exclude all transitive dependencies
+                dependencies.clear();
             });
         });
     }
